// src/generated/prisma/models/Chunk.ts
export type ChunkWhereInput = {
  AND?: Prisma.ChunkWhereInput | Prisma.ChunkWhereInput[]
  OR?: Prisma.ChunkWhereInput[]
  NOT?: Prisma.ChunkWhereInput | Prisma.ChunkWhereInput[]
  id?: Prisma.StringFilter<"Chunk"> | string
  source?: Prisma.EnumChunkSourceFilter<"Chunk"> | $Enums.ChunkSource
  pageId?: Prisma.StringNullableFilter<"Chunk"> | string | null
  materialId?: Prisma.StringNullableFilter<"Chunk"> | string | null
  ord?: Prisma.IntFilter<"Chunk"> | number
  text?: Prisma.StringFilter<"Chunk"> | string
  vector?: Prisma.BytesFilter<"Chunk"> | runtime.Bytes
  hash?: Prisma.StringFilter<"Chunk"> | string
  model?: Prisma.StringFilter<"Chunk"> | string
  createdAt?: Prisma.DateTimeFilter<"Chunk"> | Date | string
  page?: Prisma.XOR<Prisma.PageNullableScalarRelationFilter, Prisma.PageWhereInput> | null
  material?: Prisma.XOR<Prisma.MaterialNullableScalarRelationFilter, Prisma.MaterialWhereInput> | null
}